wait on init until USB is ready
The idea is: if the host is suspended, usbInterruptIsReady() will become
false and the main loop will block. The watchdog will kick in and reset
the board. The code will start up again and block again in the main
loop... repeat until the host resumes.
This patch waits after triggering USB initialization until the first
packet is received on the bus and only then starts up the main loop.