"dim" info LED in main loop
only switch on the LED while waiting for USB to become idle. In normal
operation this is always the case, so the LED will be very dim. Only
with the host suspended or USB blocked, the LED will light up until the
watchdog kicks in. This shows the state of the running sketch in a
pretty useful way.